initial object
