+++ src/input/cursor.c
@@ -16,14 +16,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-/** Low-level API to manage pointer and pointer device
- *
- * @author Dwi Asmoro Bangun
- * @copyright 2024
- * @license GPLv3
- * @inputmodule cwc.pointer
- */
-
 #include <drm_fourcc.h>
 #include <hyprcursor/hyprcursor.h>
 #include <lauxlib.h>
@@ -63,6 +55,7 @@
 #include "cwc/layout/container.h"
 #include "cwc/layout/master.h"
 #include "cwc/luaclass.h"
+#include "cwc/luaobject.h"
 #include "cwc/server.h"
 #include "cwc/signal.h"
 #include "cwc/util.h"
@@ -1148,6 +1141,9 @@ struct cwc_cursor *cwc_cursor_create(struct wlr_seat *seat)
     snprintf(size, sizeof(size) - 1, "%u", cursor->info.size);
     setenv("XCURSOR_SIZE", size, true);
 
+    lua_State *L = g_config_get_lua_State();
+    luaC_object_pointer_register(L, cursor);
+
     return cursor;
 }
 
@@ -1155,6 +1151,9 @@ static void hyprcursor_buffer_fini(struct cwc_cursor *cursor);
 
 void cwc_cursor_destroy(struct cwc_cursor *cursor)
 {
+    lua_State *L = g_config_get_lua_State();
+    luaC_object_unregister(L, cursor);
+
     // clean hyprcursor leftover
     if (cursor->images != NULL)
         hyprcursor_cursor_image_data_free(cursor->images, cursor->images_count);
@@ -1487,255 +1486,4 @@ void cleanup_pointer(struct cwc_input_manager *input_mgr)
     wl_list_remove(&input_mgr->new_vpointer_l.link);
 
     wl_list_remove(&input_mgr->request_set_shape_l.link);
-}
-
-//============= LUA ===============
-
-/** Register a mouse binding.
- *
- * @staticfct bind
- * @tparam table|number modifier Table of modifier or modifier bitfield
- * @tparam number mouse_btn Button from linux input-event-codes
- * @tparam func on_press Function to execute when pressed
- * @tparam[opt] func on_release Function to execute when released
- * @tparam[opt] table data Additional data
- * @tparam[opt] string data.group Keybinding group
- * @tparam[opt] string data.description Keybinding description
- * @noreturn
- * @see cuteful.enum.modifier
- * @see cuteful.enum.mouse_btn
- * @see cwc.kbd.bind
- */
-static int luaC_pointer_bind(lua_State *L)
-{
-    uint32_t button = luaL_checknumber(L, 2);
-    luaL_checktype(L, 3, LUA_TFUNCTION);
-
-    uint32_t modifiers = 0;
-    if (lua_istable(L, 1)) {
-        int len = lua_objlen(L, 1);
-
-        for (int i = 0; i < len; ++i) {
-            lua_rawgeti(L, 1, i + 1);
-            modifiers |= luaL_checkint(L, -1);
-        }
-
-    } else if (lua_isnumber(L, 1)) {
-        modifiers = lua_tonumber(L, 1);
-    } else {
-        luaL_error(L,
-                   "modifiers only accept array of number or modifier bitmask");
-    }
-
-    bool on_press_is_function   = lua_isfunction(L, 3);
-    bool on_release_is_function = lua_isfunction(L, 4);
-
-    if (!on_press_is_function && !on_release_is_function) {
-        luaL_error(L, "callback function is not provided");
-        return 0;
-    }
-
-    struct cwc_keybind_info info = {0};
-    info.type                    = CWC_KEYBIND_TYPE_LUA;
-
-    if (on_press_is_function) {
-        lua_pushvalue(L, 3);
-        info.luaref_press = luaL_ref(L, LUA_REGISTRYINDEX);
-    }
-
-    if (on_release_is_function) {
-        lua_pushvalue(L, 4);
-        info.luaref_release = luaL_ref(L, LUA_REGISTRYINDEX);
-    }
-
-    keybind_mouse_register(server.main_mouse_kmap, modifiers, button, info);
-
-    return 0;
-}
-
-/** Clear all mouse binding.
- *
- * @staticfct clear
- * @noreturn
- */
-static int luaC_pointer_clear(lua_State *L)
-{
-    cwc_keybind_map_clear(server.main_mouse_kmap);
-    return 0;
-}
-
-/** Get main seat pointer position.
- *
- * @staticfct get_position
- * @treturn table Pointer coords with structure {x,y}
- */
-static int luaC_pointer_get_position(lua_State *L)
-{
-    double x = server.seat->cursor->wlr_cursor->x;
-    double y = server.seat->cursor->wlr_cursor->y;
-
-    lua_createtable(L, 0, 2);
-    lua_pushnumber(L, x);
-    lua_setfield(L, -2, "x");
-    lua_pushnumber(L, y);
-    lua_setfield(L, -2, "y");
-
-    return 1;
-}
-
-/** Set main seat pointer position.
- *
- * @staticfct set_position
- * @tparam integer x
- * @tparam integer y
- * @noreturn
- */
-static int luaC_pointer_set_position(lua_State *L)
-{
-    int x = luaL_checkint(L, 1);
-    int y = luaL_checkint(L, 2);
-
-    wlr_cursor_warp(server.seat->cursor->wlr_cursor, NULL, x, y);
-
-    return 1;
-}
-
-/** Start interactive move for client under the cursor.
- *
- * @staticfct move_interactive
- * @noreturn
- */
-static int luaC_pointer_move_interactive(lua_State *L)
-{
-    start_interactive_move(NULL);
-    return 0;
-}
-
-/** Start interactive resize for client under the cursor.
- *
- * @staticfct resize_interactive
- * @noreturn
- */
-static int luaC_pointer_resize_interactive(lua_State *L)
-{
-    start_interactive_resize(NULL, 0);
-    return 0;
-}
-
-/** Stop interactive mode.
- *
- * @staticfct stop_interactive
- * @noreturn
- */
-static int luaC_pointer_stop_interactive(lua_State *L)
-{
-    stop_interactive(NULL);
-    return 0;
-}
-
-/** Set cursor size.
- *
- * @tfield integer cursor_size
- */
-static int luaC_pointer_get_cursor_size(lua_State *L)
-{
-    lua_pushnumber(L, g_config.cursor_size);
-    return 1;
-}
-static int luaC_pointer_set_cursor_size(lua_State *L)
-{
-    int size = luaL_checkint(L, 1);
-
-    g_config.cursor_size = size;
-    return 0;
-}
-
-/** Set a timeout in seconds to automatically hide cursor, set timeout to 0 to
- * disable.
- *
- * @tfield integer inactive_timeout
- */
-static int luaC_pointer_get_inactive_timeout(lua_State *L)
-{
-    lua_pushnumber(L, g_config.cursor_inactive_timeout_ms);
-    return 1;
-}
-static int luaC_pointer_set_inactive_timeout(lua_State *L)
-{
-    int seconds = luaL_checkint(L, 1);
-
-    g_config.cursor_inactive_timeout_ms = seconds * 1000;
-    return 0;
-}
-
-/** Set a threshold distance for applying common tile position in pixel unit.
- *
- * @tfield integer edge_threshold
- */
-static int luaC_pointer_get_edge_threshold(lua_State *L)
-{
-    lua_pushnumber(L, g_config.cursor_edge_threshold);
-    return 1;
-}
-static int luaC_pointer_set_edge_threshold(lua_State *L)
-{
-    int threshold = luaL_checkint(L, 1);
-
-    g_config.cursor_edge_threshold = threshold;
-    return 0;
-}
-
-/** Set color of the overlay when performing edge snapping.
- *
- * `gears.color` is not gonna work because the overlay isn't a cairo surface.
- *
- * @configfct set_edge_snapping_overlay_color
- * @tparam number red Value of red.
- * @tparam number green Value of green.
- * @tparam number blue Value of blue.
- * @tparam number alpha Alpha value.
- * @noreturn
- */
-static int luaC_pointer_set_edge_snapping_overlay_color(lua_State *L)
-{
-    float red   = luaL_checknumber(L, 1);
-    float green = luaL_checknumber(L, 2);
-    float blue  = luaL_checknumber(L, 3);
-    float alpha = luaL_checknumber(L, 4);
-
-    g_config.cursor_edge_snapping_overlay_color[0] = red;
-    g_config.cursor_edge_snapping_overlay_color[1] = green;
-    g_config.cursor_edge_snapping_overlay_color[2] = blue;
-    g_config.cursor_edge_snapping_overlay_color[3] = alpha;
-    return 0;
-}
-
-#define FIELD_RO(name)     {"get_" #name, luaC_pointer_get_##name}
-#define FIELD_SETTER(name) {"set_" #name, luaC_pointer_set_##name}
-#define FIELD(name)        FIELD_RO(name), FIELD_SETTER(name)
-
-void luaC_pointer_setup(lua_State *L)
-{
-    luaL_Reg pointer_staticlibs[] = {
-        {"bind",                            luaC_pointer_bind              },
-        {"clear",                           luaC_pointer_clear             },
-
-        {"get_position",                    luaC_pointer_get_position      },
-        {"set_position",                    luaC_pointer_set_position      },
-
-        {"move_interactive",                luaC_pointer_move_interactive  },
-        {"resize_interactive",              luaC_pointer_resize_interactive},
-        {"stop_interactive",                luaC_pointer_stop_interactive  },
-
-        FIELD(cursor_size),
-        FIELD(inactive_timeout),
-        FIELD(edge_threshold),
-        {"set_edge_snapping_overlay_color",
-         luaC_pointer_set_edge_snapping_overlay_color                      },
-
-        {NULL,                              NULL                           },
-    };
-
-    luaC_register_table(L, "cwc.pointer", pointer_staticlibs, NULL);
-    lua_setfield(L, -2, "pointer");
 }
